requirements.txt
python-telegram-bot
requests
pyclean